requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
Pillow>=10.0.0
edge-tts>=6.1.0
moviepy>=2.0.0,<3
